version 0.21.165
se han corregido modelos migratorios de la base de datos
se han corregido los enrrutadores y sus cabezeras
se ha creado la logica front end del projecto

+++ database/migrations/2023_03_23_234519_create_cars_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('cars', function (Blueprint $table) {
             $table->id();
-            $table->float('sub_total');
-            $table->float('total_purchase');
+            $table->double('sub_total');
+            $table->double('total_purchase');
             $table->softDeletes();
             $table->timestamps();
         });
